Agrega administracion de usuarios

+++ GestorArchivos/Models/Archivo.cs
@@ -12,7 +12,7 @@ namespace GestorArchivos.Models
         public int idArchivo { get; set; }
         public string nombre { get; set; }
         public string extension { get; set; }
-        public double tamanio { get; set; }
+        public decimal tamanio { get; set; }
         public string ubicacion { get; set; }
         public int idUsuario { get; set; }
